Move areDOMRectsEqual
